Fan service redemption (#35)
* removed point system

I believe ive removed the points system almost entirely, as well as correcting the spacing.

* removed unnecessary contract

FanToken was going to be used as a template for custom fan tokens... but thats not going to be implemented the way i had planned.

* basic update plan.

when calling the contract, check if there is a newer version; and update this value to the newer contract.

* Create Testing Token

Just a quick testing token so that we can mint, and test the FanService items.

* an easier way to find what pools an artist owns

just making it a bit easier for frontend!

* removed OnlyMinter

minter is now just the owner. and controllerShare is now just a constant small number.

* BUNCH OF STUFF FOR TESTING

* Update CreatorCollections.test.ts

* Adding back onlyMinter and removing MintFee(now just price)

well... i was very wrong, onlyMinter is absolutly needed and IDK how i missed that it was so vital. Also, MintFee is useless as we dont use points anymore...

* NFT system updated to genuinely be NF

* changes to testing. Renaming NEEDED

* Renamed variables

Pyro asked to rename some variables for testing

* tests completed

fixed the tests to work with new Collections system

* added tests

added some tests to throw errors, next up, not throwing errors!

* optimised some functions and removed some stuff.

removing some expensive fat, trying to make our functions cheaper to run

* tests now work for fanservice redemption.

6/6 passing tests!

* added more tests

with more tests, i should be ready to merge

* extra tests to get stats on createCard

Co-authored-by: SanchitGogia <frl.sanchit@gmail.com>