Upload.tsx, card.tsx, Discovery.tsx update
Added new input name Alt text for image/post. On the feed page when I try to query alttext with other fields it doesnt work yet. For now added name prop to alt-data, so it will display name of the post if the image has a problem.
Will fix this problem in the next commit, dont merge this yet.