init commit
The new frontend design will be temp build and implemented here. For now it uses the old codebase and is in pure react. When the new coldbase repo is out there and getting some body, I will connect the new design to it, and also make adjustments so it works with nextjs.